Update README.md
 a perl 2D shooter game using sdl perl game engine.

game to consist of only one shooter with a gun who is targeting an a flying bird moving from left to right corner of the panel. 

speed and accuracy of the bullet from the gun is a factor to shoot the bird. 

 a timer is included such that the bird will fly from left to right 10 times in 10 seconds. 

No of times the bird is shot down in 10 seconds determines winning or loosing.